Refactored background views out of mage
