+++ characters/forms/mage/advancement.py
@@ -76,7 +76,7 @@ class MageAdvancementForm(AdvancementForm):
             ("Arete", "Arete"),
             ("Quintessence", "Quintessence"),
         ]
-        if self.instance.freebies < 4:
+        if self.instance.freebies < 4 or self.instance.arete == 3:
             ADDITIONAL_CATS = [x for x in ADDITIONAL_CATS if x[0] != "Arete"]
         if self.instance.freebies < 7:
             ADDITIONAL_CATS = [x for x in ADDITIONAL_CATS if x[0] != "Sphere"]
