+++ characters/templates/characters/mage/mage/mage_freebies_form.html
@@ -103,6 +103,8 @@
         const resElement = document.getElementById("resonance_wrap");
         const noteElement = document.getElementById("note_wrap");
         const pooled_wrap = document.getElementById("pooled_wrap");
+
+        var isGroupMember = {{ object.is_group_member|lower }};
     
         categorySelectMenu.addEventListener("change", function() {
             const selectedValue = this.value;
@@ -139,8 +141,10 @@
                 exampleElement.classList.remove("invisible");
                 noteElement.classList.remove("invisible");
                 valueElement.classList.add("invisible");
-                if({{object.is_group_member}}){
+                if (isGroupMember) {
                     pooled_wrap.classList.remove("invisible");
+                } else {
+                    pooled_wrap.classList.add("invisible");
                 }
             }
             if(this.value == "Arete"){
