+++ characters/views/mage/mage.py
@@ -38,6 +38,15 @@ from characters.views.core.human import (
     HumanCharacterCreationView,
     HumanDetailView,
 )
+from characters.views.mage.background_views import (
+    MtAAlliesView,
+    MtAEnhancementView,
+    MtAFamiliarView,
+    MtALibraryView,
+    MtANodeView,
+    MtASanctumView,
+    MtAWonderView,
+)
 from characters.views.mage.mtahuman import MtAHumanAbilityView
 from core.forms.language import HumanLanguageForm
 from core.models import Language
@@ -531,7 +540,7 @@ class MageBasicsView(LoginRequiredMixin, CreateView):
             {"placeholder": "Enter concept here"}
         )
         form.fields["image"].required = False
-        if not self.request.user.profile.mta_st:
+        if not self.request.user.profile.is_st():
             form.fields["affiliation"].queryset = form.fields[
                 "affiliation"
             ].queryset.exclude(name__in=["Nephandi", "Marauders"])
@@ -983,7 +992,7 @@ class MageFreebiesView(SpecialUserMixin, UpdateView):
             trait = Background.objects.get(pk=form.data["example"])
             cost *= trait.multiplier
             value = 1
-            if form.data["pooled"] == "on":
+            if "pooled" in form.data.keys():
                 pbgr = PooledBackgroundRating.objects.get_or_create(
                     bg=trait, group=self.object.get_group(), note=form.data["note"]
                 )[0]
@@ -1219,756 +1228,32 @@ class MageRoteView(SpecialUserMixin, CreateView):
         return super().form_invalid(form)
 
 
-class MageNodeView(SpecialUserMixin, MultipleFormsetsMixin, FormView):
-    form_class = NodeForm
-    template_name = "characters/mage/mage/chargen.html"
-    formsets = {
-        "rz_form": RealityZonePracticeRatingFormSet,
-        "resonance_form": NodeResonancePracticeRatingFormSet,
-        "mf_form": NodeMeritFlawFormSet,
-    }
-
-    def form_valid(self, form):
-        context = self.get_context_data()
-        mage = context["object"]
-
-        n = Node(
-            name=form.cleaned_data["name"],
-            description=form.cleaned_data["description"],
-            ratio=form.cleaned_data["ratio"],
-            size=form.cleaned_data["size"],
-            quintessence_form=form.cleaned_data["quintessence_form"],
-            tass_form=form.cleaned_data["tass_form"],
-            gauntlet=3,
-            owned_by=mage,
-            chronicle=mage.chronicle,
-            owner=mage.owner,
-            status="Sub",
-        )
-        n.set_rank(self.current_node.rating)
-        n.points -= form.cleaned_data["ratio"]
-        n.points -= form.cleaned_data["size"]
-
-        resonance_data = self.get_form_data("resonance_form")
-        for res in resonance_data:
-            res["resonance"] = Resonance.objects.get_or_create(name=res["resonance"])[0]
-            res["rating"] = int(res["rating"])
-            if res["rating"] > 5:
-                form.add_error(
-                    None,
-                    "Resonance may not be higher than 5",
-                )
-                return self.form_invalid(form)
-        total_resonance = sum([x["rating"] for x in resonance_data])
-
-        mf_data = self.get_form_data("mf_form")
-        for res in mf_data:
-            res["mf"] = MeritFlaw.objects.get(id=res["mf"])
-            res["rating"] = int(res["rating"])
-        total_mfs = sum([x["rating"] for x in mf_data])
-
-        rz_data = self.get_form_data("rz_form")
-        for res in rz_data:
-            res["practice"] = Practice.objects.get(id=res["practice"])
-            res["rating"] = int(res["rating"])
-        total_rz = sum([x["rating"] for x in rz_data])
-        total_positive = sum([x["rating"] for x in rz_data if x["rating"] > 0])
-        if total_rz != 0:
-            form.add_error(None, "Ratings must total 0")
-            return super().form_invalid(form)
-        if total_positive != self.current_node.rating:
-            form.add_error(None, "Positive Ratings must equal Node rating")
-            return super().form_invalid(form)
-
-        if total_resonance < n.rank:
-            form.add_error(None, "Resonance total must match or exceed rank")
-            return self.form_invalid(form)
-        n.points -= total_resonance - n.rank
-        n.points -= total_mfs
-        if n.points < 0:
-            form.add_error(
-                None,
-                "Node invalid: spend fewer points on merits/flaws, resonance, size, or ratio",
-            )
-            return self.form_invalid(form)
-        n.update_output()
-        n.save()
-
-        self.current_node.note = n.name
-        self.current_node.url = n.get_absolute_url()
-        self.current_node.complete = True
-        self.current_node.save()
-
-        for resonance in resonance_data:
-            NodeResonanceRating.objects.create(
-                node=n, resonance=resonance["resonance"], rating=resonance["rating"]
-            )
-
-        for mf in mf_data:
-            NodeMeritFlawRating.objects.create(node=n, mf=mf["mf"], rating=mf["rating"])
-
-        rzone = RealityZone.objects.create(name="{n.name} Reality Zone")
-        n.reality_zone = rzone
-        n.save()
-        for rz in rz_data:
-            ZoneRating.objects.create(
-                zone=rzone, practice=rz["practice"], rating=rz["rating"]
-            )
-
-        if (
-            BackgroundRating.objects.filter(
-                char=mage,
-                bg=Background.objects.get(property_name="node"),
-                complete=False,
-            ).count()
-            == 0
-        ):
-            mage.creation_status += 1
-            mage.save()
-            for step in [
-                "library",
-                "familiar",
-                "wonder",
-                "enhancement",
-                "sanctum",
-                "allies",
-            ]:
-                if (
-                    BackgroundRating.objects.filter(
-                        bg=Background.objects.get(property_name=step),
-                        char=mage,
-                        complete=False,
-                    ).count()
-                    == 0
-                ):
-                    mage.creation_status += 1
-                else:
-                    mage.save()
-                    break
-            mage.save()
-        return HttpResponseRedirect(mage.get_absolute_url())
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["object"] = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        context["is_approved_user"] = self.check_if_special_user(
-            context["object"], self.request.user
-        )
-        context["points"] = 3 * self.current_node.rating
-        context["current_node"] = self.current_node
-        return context
-
-    def get_form(self, form_class=None):
-        mage = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        self.current_node = BackgroundRating.objects.filter(
-            char=mage, bg=Background.objects.get(property_name="node"), complete=False
-        ).first()
-        form = super().get_form(form_class)
-        form.fields["name"].initial = self.current_node.note
-        form.fields["quintessence_form"].widget.attrs.update(
-            {"placeholder": "Enter quintessence form here"}
-        )
-        form.fields["tass_form"].widget.attrs.update(
-            {"placeholder": "Enter tass form here"}
-        )
-        form.fields["description"].widget.attrs.update(
-            {"placeholder": "Enter description here"}
-        )
-        return form
-
-
-class MageLibraryView(SpecialUserMixin, CreateView):
-    model = Library
-    fields = ["name", "description", "parent"]
+class MageAlliesView(MtAAlliesView):
     template_name = "characters/mage/mage/chargen.html"
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["object"] = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        context["is_approved_user"] = self.check_if_special_user(
-            context["object"], self.request.user
-        )
-        context["current_library"] = self.current_library
-        return context
 
-    def form_valid(self, form):
-        context = self.get_context_data()
-        mage = context["object"]
-        l = Library(
-            **form.cleaned_data,
-            owned_by=mage,
-            owner=mage.owner,
-            chronicle=mage.chronicle,
-            faction=mage.faction,
-            rank=self.current_library.rating,
-            status="Sub",
-        )
-        l.save()
-        for _ in range(l.rank):
-            l.random_book()
-
-        self.current_library.note = l.name
-        self.current_library.url = l.get_absolute_url()
-        self.current_library.complete = True
-        self.current_library.save()
-
-        if (
-            BackgroundRating.objects.filter(
-                char=mage,
-                bg=Background.objects.get(property_name="library"),
-                complete=False,
-            ).count()
-            == 0
-        ):
-            mage.creation_status += 1
-            mage.save()
-            for step in [
-                "familiar",
-                "wonder",
-                "enhancement",
-                "sanctum",
-                "allies",
-            ]:
-                if (
-                    BackgroundRating.objects.filter(
-                        bg=Background.objects.get(property_name=step),
-                        char=mage,
-                        complete=False,
-                    ).count()
-                    == 0
-                ):
-                    mage.creation_status += 1
-                else:
-                    mage.save()
-                    break
-            mage.save()
-        return HttpResponseRedirect(mage.get_absolute_url())
-
-    def get_form(self, form_class=None):
-        form = super().get_form(form_class)
-        obj = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        self.current_library = BackgroundRating.objects.filter(
-            char=obj, bg=Background.objects.get(property_name="library"), complete=False
-        ).first()
-        form.fields["name"].initial = (
-            self.current_library.note or f"{obj.name}'s Library"
-        )
-        form.fields["parent"].empty_label = "Choose a Parent Location"
-        form.fields["description"].widget.attrs.update(
-            {"placeholder": "Enter description here"}
-        )
-        return form
-
-
-class MageFamiliarView(SpecialUserMixin, FormView):
-    form_class = FamiliarForm
+class MageEnhancementView(MtAEnhancementView):
     template_name = "characters/mage/mage/chargen.html"
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["object"] = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        context["is_approved_user"] = self.check_if_special_user(
-            context["object"], self.request.user
-        )
-        return context
 
-    def form_invalid(self, form):
-        errors = form.errors
-        if "faction" in errors:
-            del errors["faction"]
-        if "subfaction" in errors:
-            del errors["subfaction"]
-
-        if not errors:
-            return self.form_valid(form)
-        return super().form_invalid(form)
-
-    def form_valid(self, form):
-        context = self.get_context_data()
-        mage = context["object"]
-
-        if form.data["faction"]:
-            faction = MageFaction.objects.get(pk=form.data["faction"])
-        else:
-            faction = None
-        if form.data["subfaction"]:
-            subfaction = MageFaction.objects.get(pk=form.data["subfaction"])
-        else:
-            subfaction = None
-
-        c = Companion(
-            name=form.cleaned_data["name"],
-            nature=form.cleaned_data["nature"],
-            demeanor=form.cleaned_data["demeanor"],
-            affiliation=form.cleaned_data["affiliation"],
-            faction=faction,
-            subfaction=subfaction,
-            companion_type="familiar",
-            concept=form.cleaned_data["concept"],
-            chronicle=mage.chronicle,
-            owner=mage.owner,
-            npc=True,
-            companion_of=mage,
-        )
-        x = BackgroundRating.objects.filter(
-            char=mage,
-            bg=Background.objects.get(property_name="familiar"),
-            complete=False,
-        ).first()
-        c.freebies = 10 * x.rating
-        c.save()
-        x.url = c.get_absolute_url()
-        x.complete = True
-        x.note = c.name
-        x.save()
-        if (
-            BackgroundRating.objects.filter(
-                char=mage,
-                bg=Background.objects.get(property_name="familiar"),
-                complete=False,
-            ).count()
-            == 0
-        ):
-            mage.creation_status += 1
-            mage.save()
-            for step in [
-                "wonder",
-                "enhancement",
-                "sanctum",
-                "allies",
-            ]:
-                if (
-                    BackgroundRating.objects.filter(
-                        bg=Background.objects.get(property_name=step),
-                        char=mage,
-                        complete=False,
-                    ).count()
-                    == 0
-                ):
-                    mage.creation_status += 1
-                else:
-                    mage.save()
-                    break
-            mage.save()
-        return HttpResponseRedirect(mage.get_absolute_url())
-
-    def get_form(self, form_class=None):
-        form = super().get_form(form_class)
-        form.fields["name"].initial = (
-            BackgroundRating.objects.filter(
-                char=get_object_or_404(Human, pk=self.kwargs.get("pk")),
-                bg=Background.objects.get(property_name="familiar"),
-                complete=False,
-            )
-            .first()
-            .note
-        )
-        return form
-
-
-class MageWonderView(SpecialUserMixin, MultipleFormsetsMixin, FormView):
-    form_class = WonderForm
-    formsets = {
-        "effects_form": EffectFormSet,
-        "resonance_form": WonderResonancePracticeRatingFormSet,
-    }
+class MageFamiliarView(MtAFamiliarView):
     template_name = "characters/mage/mage/chargen.html"
 
-    wonder_classes = {
-        "charm": Charm,
-        "artifact": Artifact,
-        "talisman": Talisman,
-    }
 
-    def get_context_data(self, **kwargs) -> dict[str, Any]:
-        context = super().get_context_data(**kwargs)
-        context["object"] = Mage.objects.get(id=self.kwargs["pk"])
-        context["is_approved_user"] = self.check_if_special_user(
-            context["object"], self.request.user
-        )
-        context["points"] = 3 * context["object"].wonder
-        context["current_wonder"] = self.current_wonder
-        return context
-
-    def form_valid(self, form):
-        context = self.get_context_data()
-        mage = context["object"]
-        if form.cleaned_data["select_or_create_wonder"]:
-            del form.cleaned_data["select_or_create_wonder"]
-            del form.cleaned_data["wonder_options"]
-            if form.cleaned_data["wonder_type"] == "artifact":
-                del form.cleaned_data["arete"]
-            wonder_type = form.cleaned_data["wonder_type"]
-            del form.cleaned_data["wonder_type"]
-            w = self.wonder_classes[wonder_type](
-                **form.cleaned_data,
-                rank=self.current_wonder.rating,
-                chronicle=mage.chronicle,
-                owner=mage.owner,
-                status="Sub",
-            )
-            w.owned_by.add(mage)
-            points = 3 * w.rank
-
-            resonance_data = self.get_form_data("resonance_form")
-            for res in resonance_data:
-                res["resonance"] = Resonance.objects.get_or_create(
-                    name=res["resonance"]
-                )[0]
-                res["rating"] = int(res["rating"])
-                if res["rating"] > 5:
-                    form.add_error(
-                        None,
-                        "Resonance may not be higher than 5",
-                    )
-                    return self.form_invalid(form)
-            total_resonance = sum([x["rating"] for x in resonance_data])
-            if total_resonance < w.rank:
-                form.add_error(None, "Resonance must be at least rank")
-                return self.form_invalid(form)
-            if wonder_type == "charm":
-                max_cost = w.rank
-            else:
-                max_cost = 2 * w.rank
-
-            effects = []
-            total_effect_cost = 0
-            effects_data = self.get_form_data("effects_form")
-            if wonder_type == "charm" and len(effects_data) > 1:
-                form.add_error(None, "Charms can only have one power")
-                return self.form_invalid(form)
-            elif wonder_type == "artifact" and len(effects_data) > 1:
-                form.add_error(None, "Artifacts can only have one power")
-                return self.form_invalid(form)
-            elif wonder_type == "talisman" and len(effects_data) > w.rank:
-                form.add_error(None, "Talismans may up to their rank in effects")
-                return self.form_invalid(form)
-            for effect in effects_data:
-                for sphere in Sphere.objects.all():
-                    effect[sphere.property_name] = int(effect[sphere.property_name])
-                e = Effect(**effect)
-                effects.append(e)
-                total_effect_cost += e.cost()
-                if total_effect_cost > max_cost:
-                    form.add_error(
-                        None,
-                        "Effects cost more than allowed: rank for Charms, twice rank for Artifacts and Talismans",
-                    )
-                    return self.form_invalid(form)
-            cost = (total_resonance - w.rank) + total_effect_cost
-            if wonder_type != "artifact":
-                cost += w.arete - w.rank
-            if cost > points:
-                form.add_error(
-                    None,
-                    "Extra Resonance, Arete, and Effects must be less than 3 times the rank of the Wonder",
-                )
-                return self.form_invalid(form)
-
-            w.save()
-            for e in effects:
-                e.save()
-                if wonder_type in ["charm", "artifact"]:
-                    w.power = e
-                else:
-                    w.powers.add(e)
-            w.save()
-
-            for resonance in resonance_data:
-                WonderResonanceRating.objects.create(
-                    wonder=None,
-                    resonance=resonance["resonance"],
-                    rating=resonance["rating"],
-                )
-        else:
-            w = form.cleaned_data["wonder_options"]
-            w.owned_by.add(mage)
-            w.save()
-
-        self.current_wonder.note = w.name
-        self.current_wonder.url = w.get_absolute_url()
-        self.current_wonder.complete = False
-        self.current_wonder.save()
-
-        if (
-            BackgroundRating.objects.filter(
-                char=mage,
-                bg=Background.objects.get(property_name="wonder"),
-                complete=False,
-            ).count()
-            == 0
-        ):
-            mage.creation_status += 1
-            mage.save()
-            for step in [
-                "enhancement",
-                "sanctum",
-                "allies",
-            ]:
-                if (
-                    BackgroundRating.objects.filter(
-                        bg=Background.objects.get(property_name=step),
-                        char=mage,
-                        complete=False,
-                    ).count()
-                    == 0
-                ):
-                    mage.creation_status += 1
-                else:
-                    mage.save()
-                    break
-            mage.save()
-        return HttpResponseRedirect(mage.get_absolute_url())
-
-    def get_form(self, form_class=None):
-        form = super().get_form(form_class)
-        obj = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        self.current_wonder = BackgroundRating.objects.filter(
-            char=obj, bg=Background.objects.get(property_name="wonder"), complete=False
-        ).first()
-        form.fields["name"].initial = self.current_wonder.note
-        form.fields["description"].widget.attrs.update(
-            {"placeholder": "Enter description here"}
-        )
-        return form
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        mage_id = self.kwargs.get("pk")
-        mage = Mage.objects.get(pk=mage_id)
-        kwargs["instance"] = mage
-        bgr = BackgroundRating.objects.filter(
-            char=mage, bg=Background.objects.get(property_name="wonder"), complete=False
-        ).first()
-        kwargs["rank"] = bgr.rating
-        return kwargs
-
-
-class MageEnhancementView(SpecialUserMixin, FormView):
-    form_class = EnhancementForm
+class MageLibraryView(MtALibraryView):
     template_name = "characters/mage/mage/chargen.html"
 
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        mage = Mage.objects.get(id=self.kwargs["pk"])
-        self.current_enhancement = BackgroundRating.objects.filter(
-            char=mage,
-            bg=Background.objects.get(property_name="enhancement"),
-            complete=False,
-        ).first()
-        kwargs["rank"] = self.current_enhancement.rating
-        return kwargs
 
-    def get_form(self, form_class=None):
-        form = super().get_form(form_class)
-        obj = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        self.current_enhancement = BackgroundRating.objects.filter(
-            char=obj,
-            bg=Background.objects.get(property_name="enhancement"),
-            complete=False,
-        ).first()
-        return form
-
-    def get_context_data(self, **kwargs) -> dict[str, Any]:
-        context = super().get_context_data(**kwargs)
-        context["object"] = Mage.objects.get(id=self.kwargs["pk"])
-        context["is_approved_user"] = self.check_if_special_user(
-            context["object"], self.request.user
-        )
-        context["current_enhancement"] = self.current_enhancement
-        return context
-
-    def form_valid(self, form):
-        context = self.get_context_data()
-        mage = context["object"]
-        # Check Data Valid
-        if form.save(mage=mage):
-            if (
-                BackgroundRating.objects.filter(
-                    char=mage,
-                    bg=Background.objects.get(property_name="enhancement"),
-                    complete=False,
-                ).count()
-                == 0
-            ):
-                mage.creation_status += 1
-                mage.save()
-                for step in [
-                    "sanctum",
-                    "allies",
-                ]:
-                    if (
-                        BackgroundRating.objects.filter(
-                            bg=Background.objects.get(property_name=step),
-                            char=mage,
-                            complete=False,
-                        ).count()
-                        == 0
-                    ):
-                        mage.creation_status += 1
-                    else:
-                        mage.save()
-                        break
-                mage.save()
-            return HttpResponseRedirect(mage.get_absolute_url())
-        return super().form_invalid(form)
-
-
-class MageSanctumView(SpecialUserMixin, CreateView):
-    model = Sanctum
-    form_class = SanctumForm
+class MageNodeView(MtANodeView):
     template_name = "characters/mage/mage/chargen.html"
 
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        mage_id = self.kwargs.get("pk")
-        context["object"] = Mage.objects.get(id=mage_id)
-        context["rz_form"] = RealityZonePracticeRatingFormSet()
-        context["form"].fields["name"].initial = f"{context['object']}'s Sanctum"
-        context["is_approved_user"] = self.check_if_special_user(
-            context["object"], self.request.user
-        )
-        context["current_sanctum"] = self.current_sanctum
-        return context
 
-    def form_valid(self, form):
-        context = self.get_context_data()
-        mage = context["object"]
-        rz_form = context["rz_form"]
-        practices = [
-            v
-            for k, v in form.data.items()
-            if k.startswith("zonerating_set-") and k.endswith("-practice")
-        ]
-        ratings = [
-            int(v)
-            for k, v in form.data.items()
-            if k.startswith("zonerating_set-") and k.endswith("-rating")
-        ]
-        pairs = [
-            (Practice.objects.get(id=practice), rating)
-            for practice, rating in zip(practices, ratings)
-            if rating != 0
-        ]
-        total_rating = sum([v for k, v in pairs])
-        total_positive = sum([v for k, v in pairs if v > 0])
-        if total_rating != 0:
-            form.add_error(None, "Ratings must total 0")
-            return super().form_invalid(form)
-        if total_positive != self.current_sanctum.rating:
-            form.add_error(None, "Positive Ratings must equal Sanctum rating")
-            return super().form_invalid(form)
-        rz = RealityZone.objects.create(name=f"{mage.name}'s Sanctum Reality Zone")
-        for p, r in pairs:
-            ZoneRating.objects.create(zone=rz, practice=p, rating=r)
-
-        s = form.save(mage, reality_zone=rz)
-
-        self.current_sanctum.note = s.name
-        self.current_sanctum.url = s.get_absolute_url()
-        self.current_sanctum.complete = False
-        self.current_sanctum.save()
-        if form.save(mage, reality_zone=rz):
-            if (
-                BackgroundRating.objects.filter(
-                    char=mage,
-                    bg=Background.objects.get(property_name="sanctum"),
-                    complete=False,
-                ).count()
-                == 0
-            ):
-                mage.creation_status += 1
-                mage.save()
-                for step in [
-                    "allies",
-                ]:
-                    if (
-                        BackgroundRating.objects.filter(
-                            bg=Background.objects.get(property_name=step),
-                            char=mage,
-                            complete=False,
-                        ).count()
-                        == 0
-                    ):
-                        mage.creation_status += 1
-                    else:
-                        mage.save()
-                        break
-                mage.save()
-            return HttpResponseRedirect(mage.get_absolute_url())
-        return super().form_invalid(form)
-
-    def get_form(self, form_class=None):
-        form = super().get_form(form_class)
-        obj = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        self.current_sanctum = BackgroundRating.objects.filter(
-            char=obj, bg=Background.objects.get(property_name="sanctum"), complete=False
-        ).first()
-        form.fields["name"].initial = self.current_sanctum.note
-        form.fields["description"].widget.attrs.update(
-            {"placeholder": "Enter description here"}
-        )
-        form.fields["parent"].queryset = LocationModel.objects.filter(
-            chronicle=obj.chronicle
-        )
-        return form
-
-
-class MageAlliesView(SpecialUserMixin, FormView):
-    form_class = AllyForm
+class MageSanctumView(MtASanctumView):
     template_name = "characters/mage/mage/chargen.html"
 
-    ally_types = {"human": MtAHuman, "mage": Mage, "spirit": SpiritCharacter}
 
-    def get_context_data(self, **kwargs) -> dict[str, Any]:
-        context = super().get_context_data(**kwargs)
-        context["object"] = Mage.objects.get(id=self.kwargs["pk"])
-        context["is_approved_user"] = self.check_if_special_user(
-            context["object"], self.request.user
-        )
-        context["current_ally"] = self.current_ally
-        return context
-
-    def form_valid(self, form):
-        context = self.get_context_data()
-        mage = context["object"]
-        char_class = self.ally_types[form.cleaned_data["ally_type"]]
-        a = char_class.objects.create(
-            name=form.cleaned_data["name"],
-            concept=form.cleaned_data["name"],
-            notes=form.cleaned_data["name"]
-            + f"<br> Rank {self.current_ally.rating} Ally for {mage.name}",
-            chronicle=mage.chronicle,
-            npc=True,
-            status="Un",
-        )
-        mage.allied_characters.add(a)
-
-        self.current_ally.note = a.name
-        self.current_ally.url = a.get_absolute_url()
-        self.current_ally.complete = False
-        self.current_ally.save()
-
-        if (
-            BackgroundRating.objects.filter(
-                char=mage,
-                bg=Background.objects.get(property_name="allies"),
-                complete=False,
-            ).count()
-            == 0
-        ):
-            mage.creation_status += 1
-            mage.save()
-        return HttpResponseRedirect(mage.get_absolute_url())
-
-    def get_form(self, form_class=None):
-        form = super().get_form(form_class)
-        obj = get_object_or_404(Human, pk=self.kwargs.get("pk"))
-        self.current_ally = BackgroundRating.objects.filter(
-            char=obj, bg=Background.objects.get(property_name="allies"), complete=False
-        ).first()
-        form.fields["name"].initial = self.current_ally.note
-        return form
+class MageWonderView(MtAWonderView):
+    template_name = "characters/mage/mage/chargen.html"
 
 
 class MageSpecialtiesView(SpecialUserMixin, FormView):
